requirements/base.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
black==25.1.0
boto3==1.38.6
botocore==1.38.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contextlib2==21.6.0
cryptography==44.0.2
decorator==5.2.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
django-storages==1.14.6
executing==2.2.0
idna==3.10
iniconfig==2.1.0
injector==0.22.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.12.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
